[DMS-1337] Update CMS resource identifier columns to INT/int to match Admin API and the draft Management API v3 spec (#1140)
* Narrow BIGINT to INT in the PostgreSQL deploy scripts

* Narrow BIGINT to INT in the SQL Server deploy scripts

* Retype DataModel response, command, and query models from long to int

* Retype backend repository interfaces, result records, and OpenIddict abstractions

* Retype the PostgreSQL repositories and fix the bigint[] array casts

* Retype the SQL Server repositories

* Retype the frontend minimal-API handlers and query models

* Fix the hard-coded int64 id parameter in the hand-built metadata document

* Assert the dmscs bigint allowlist and integer identity columns in the PostgreSQL shape test

* Assert the dmscs bigint allowlist and int identity columns in the SQL Server deploy test

* Assert int32 identifiers across both OpenAPI documents

* Add a model identifier contract test for the DTOs no other test can reach

* Add route-binding boundary tests for the narrowed id

* Add an above-int.MaxValue education-organization round-trip test to both integration suites

* Retype long literals and fakes in the existing frontend and backend unit tests

* Retype long literals in both integration suites

* Update the CMS schema documentation

* Changes after rebase

* Self code review - Test coverage

+++ src/config/backend/EdFi.DmsConfigurationService.Backend.Keycloak/KeycloakClientRepository.cs
@@ -39,7 +39,7 @@ public class KeycloakClientRepository(
         string scope,
         string namespacePrefixes,
         string educationOrganizationIds,
-        long[]? dataStoreIds = null,
+        int[]? dataStoreIds = null,
         bool isApproved = true
     )
     {
@@ -306,7 +306,7 @@ public class KeycloakClientRepository(
         string displayName,
         string scope,
         string educationOrganizationIds,
-        long[]? dataStoreIds = null,
+        int[]? dataStoreIds = null,
         bool isApproved = true,
         // Intentionally unused for Keycloak. The update preserves the client's identity, so its
         // service account and realm role mappings survive without any reassignment; the parameter
@@ -515,7 +515,7 @@ public class KeycloakClientRepository(
             protocolMappers.Add(EducationOrganizationProtocolMapper(educationOrganizationIds));
         }
 
-        void ReplaceDataStoreIdsClaim(List<ClientProtocolMapper> protocolMappers, long[]? dataStoreIds)
+        void ReplaceDataStoreIdsClaim(List<ClientProtocolMapper> protocolMappers, int[]? dataStoreIds)
         {
             protocolMappers.RemoveAll(mapper => HasClaimName(mapper, "dataStoreIds"));
