+++ src/config/datamodel/EdFi.DmsConfigurationService.DataModel/Model/ApiClient/ApiClientCredentialsResponse.cs
@@ -7,8 +7,8 @@ namespace EdFi.DmsConfigurationService.DataModel.Model.ApiClient;
 
 public class ApiClientCredentialsResponse
 {
-    public required long Id { get; set; }
-    public required long ApplicationId { get; set; }
+    public required int Id { get; set; }
+    public required int ApplicationId { get; set; }
     public required string Name { get; set; }
     public required string Key { get; set; }
     public required string Secret { get; set; }
