+++ src/config/datamodel/EdFi.DmsConfigurationService.DataModel/Model/DataStoreContext/DataStoreContextInsertCommand.cs
@@ -9,7 +9,7 @@ namespace EdFi.DmsConfigurationService.DataModel.Model.DataStoreContext;
 
 public class DataStoreContextInsertCommand
 {
-    public long DataStoreId { get; set; }
+    public int DataStoreId { get; set; }
     public string ContextKey { get; set; } = "";
     public string ContextValue { get; set; } = "";
